Remove dummy code and add event endpoint
Removes example code and adds an events post endpoint for receiving
webhooks. Currently it is not working.
I fixed a 413, due to our global body limit and now we are
receiving a 400 error. I assume the payload is encryped and or
signed and we have to load the private key for the App to decrypt it.
Currently I am trying to figure out how to log/inspect the raw
payload.