Starts event endpoint implementation
After researching a way to distinguish payloads, I learned, that we
can use a custom  X header.
Based on that we are now able to define a Push struct that mirrors
the Push event so that we can extract the commit message and
create a check-run via graphql.

Next we wait for the check-run to complete to be happy.